+++ example/config/config.exs
@@ -1,11 +1,11 @@
 # This file is responsible for configuring your application
-# and its dependencies with the aid of the Mix.Config module.
+# and its dependencies with the aid of the Config module.
 #
 # This configuration file is loaded before any dependency and
 # is restricted to this project.
 
 # General application configuration
-use Mix.Config
+import Config
 
 config :nested,
   ecto_repos: [Nested.Repo]
@@ -13,15 +13,19 @@ config :nested,
 # Configures the endpoint
 config :nested, NestedWeb.Endpoint,
   url: [host: "localhost"],
-  secret_key_base:
-    "OhEawkMknnlFF0OVrN6MCznZxybsa8/8hX2iN6oHtwucEOJmSxdBWGpPrn8hB7iw",
-  render_errors: [
-    view: NestedWeb.ErrorView,
-    accepts: ~w(html json),
-    layout: false
-  ],
+  render_errors: [view: NestedWeb.ErrorView, accepts: ~w(html json), layout: false],
   pubsub_server: Nested.PubSub,
-  live_view: [signing_salt: "YExwpGGo"]
+  live_view: [signing_salt: "mi1bglNb"]
+
+# Configure esbuild (the version is required)
+config :esbuild,
+  version: "0.14.0",
+  default: [
+    args:
+      ~w(js/app.js --bundle --target=es2017 --outdir=../priv/static/assets --external:/fonts/* --external:/images/*),
+    cd: Path.expand("../assets", __DIR__),
+    env: %{"NODE_PATH" => Path.expand("../deps", __DIR__)}
+  ]
 
 # Configures Elixir's Logger
 config :logger, :console,
@@ -33,4 +37,4 @@ config :phoenix, :json_library, Jason
 
 # Import environment specific config. This must remain at the bottom
 # of this file so it overrides the configuration defined above.
-import_config "#{Mix.env()}.exs"
+import_config "#{config_env()}.exs"